pyproject.toml
[project]
name = "velo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "jupyterlab>=4.4.7",
    "matplotlib>=3.10.6",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "ultralytics>=8.3.198",
]
